Implement exceptions for path parameter
